# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MarkHub API
  version: 0.1.0
  description: 'Self-hosted bookmark hub — unified contract for Docker (FastAPI) and
    Cloudflare Workers. Error envelope: `{ "error": { "code", "message", "details?"
    } }`.'
servers:
- url: /
paths:
...
  /api/v1/bookmarks:
    get:
      tags:
      - bookmarks
      summary: List Bookmarks
      operationId: list_bookmarks_api_v1_bookmarks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: folder_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: include_archived
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Archived
      - name: visibility
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Visibility
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - bookmarks
      summary: Create Bookmark
      operationId: create_bookmark_api_v1_bookmarks_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bookmarks/{bookmark_id}:
    get:
      tags:
      - bookmarks
      summary: Get Bookmark
      operationId: get_bookmark_api_v1_bookmarks__bookmark_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: bookmark_id
        in: path
        required: true
        schema:
          type: string
          title: Bookmark Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - bookmarks
      summary: Patch Bookmark
      operationId: patch_bookmark_api_v1_bookmarks__bookmark_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: bookmark_id
        in: path
        required: true
        schema:
          type: string
          title: Bookmark Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - bookmarks
      summary: Delete Bookmark
      operationId: delete_bookmark_api_v1_bookmarks__bookmark_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: bookmark_id
        in: path
        required: true
        schema:
          type: string
          title: Bookmark Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/bookmarks/reorder:
    post:
      tags:
      - bookmarks
      summary: Reorder
      operationId: reorder_api_v1_bookmarks_reorder_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__api__bookmarks__ReorderBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/bookmarks/batch:
    post:
      tags:
      - bookmarks
      summary: Batch
      operationId: batch_api_v1_bookmarks_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__api__bookmarks__BatchBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/settings:
    get:
      tags:
      - settings
      summary: Get Settings Api
      operationId: get_settings_api_api_v1_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    put:
      tags:
      - settings
      summary: Put Settings
      operationId: put_settings_api_v1_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/settings/ai:
    get:
      tags:
      - settings
      summary: Get Ai
      operationId: get_ai_api_v1_settings_ai_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    put:
      tags:
      - settings
      summary: Put Ai
      operationId: put_ai_api_v1_settings_ai_put
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/settings/ai/test:
    post:
      tags:
      - settings
      summary: Test Ai
      operationId: test_ai_api_v1_settings_ai_test_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/v1/settings/mcp:
    get:
      tags:
      - settings
      summary: Get Mcp
      operationId: get_mcp_api_v1_settings_mcp_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
    put:
      tags:
      - settings
      summary: Put Mcp
      operationId: put_mcp_api_v1_settings_mcp_put
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/settings/mcp/token:
    post:
      tags:
      - settings
      summary: Rotate Mcp Token
      operationId: rotate_mcp_token_api_v1_settings_mcp_token_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /api/v1/clean/jobs:
    post:
      tags:
      - clean
      summary: Create Job
      operationId: create_job_api_v1_clean_jobs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanJobBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/clean/jobs/{job_id}:
    get:
      tags:
      - clean
      summary: Get Job
      operationId: get_job_api_v1_clean_jobs__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/clean/apply:
    post:
      tags:
      - clean
      summary: Apply
      operationId: apply_api_v1_clean_apply_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/analytics/profile:
    get:
      tags:
      - clean
      summary: Profile
      operationId: profile_api_v1_analytics_profile_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /api/v1/mcp:
    post:
      tags:
      - mcp
      summary: Mcp Streamable
      description: 'MCP Streamable HTTP transport: JSON-RPC over POST /mcp.

        Accepts initialize / tools/list / tools/call. Returns JSON or SSE when Accept
        prefers it.'
      operationId: mcp_streamable_api_v1_mcp_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - mcp
      summary: Mcp Get Info
      operationId: mcp_get_info_api_v1_mcp_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApplyBody:
      properties:
        issue_ids:
          items:
            type: string
          type: array
          title: Issue Ids
        mark_link_status:
          type: boolean
          title: Mark Link Status
          default: false
      type: object
      required:
      - issue_ids
      title: ApplyBody
...
    BookmarkCreate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        url:
          type: string
          title: Url
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          anyOf:
          - type: string
          - type: 'null'
          title: Visibility
          default: private
        is_favorite:
          type: boolean
          title: Is Favorite
          default: false
        is_archived:
          type: boolean
          title: Is Archived
          default: false
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        sort_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sort Order
      type: object
      required:
      - url
      title: BookmarkCreate
    BookmarkPatch:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          anyOf:
          - type: string
          - type: 'null'
          title: Visibility
        is_favorite:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Favorite
        is_archived:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Archived
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        sort_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sort Order
        link_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Link Status
      type: object
      title: BookmarkPatch
...
    CleanJobBody:
      properties:
        check_invalid:
          type: boolean
          title: Check Invalid
          default: false
        concurrency:
          type: integer
          title: Concurrency
          default: 8
      type: object
      title: CleanJobBody
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SettingsBody:
      properties:
        theme:
          anyOf:
          - type: string
          - type: 'null'
          title: Theme
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        site_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Title
        site_icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Site Icon
        accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Accent
        wallpaper:
          anyOf:
          - type: string
          - type: 'null'
          title: Wallpaper
        card_density:
          anyOf:
          - type: string
          - type: 'null'
          title: Card Density
        root_folder_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Root Folder Id
        pinned_folder_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Pinned Folder Ids
        expanded_folder_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Expanded Folder Ids
        collection_board_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Collection Board Name
        compare_engines:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Compare Engines
        compare_active_ids:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Compare Active Ids
        discover_widgets:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Discover Widgets
        info_entries:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Info Entries
      type: object
      title: SettingsBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    app__api__bookmarks__BatchBody:
      properties:
        action:
          type: string
          title: Action
        ids:
          items:
            type: string
          type: array
          title: Ids
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
      type: object
      required:
      - action
      - ids
      title: BatchBody
    app__api__bookmarks__ReorderBody:
      properties:
        folder_id:
          type: string
          title: Folder Id
        ordered_ids:
          items:
            type: string
          type: array
          title: Ordered Ids
      type: object
      required:
      - folder_id
      - ordered_ids
      title: ReorderBody
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer